fill basic funtion functions. Also it is defined more pins

+++ BackEnd/src/hw_lib/basic_function.h
@@ -7,20 +7,24 @@
 
 #include "drivers/driver_led.h"
 #include "drivers/hardware_interface.h"
+#include "drivers/driver_stepper.h"
 
 //----------Driver varibles----------//
 extern DriverLed led_red;
 extern DriverLed buzzer;
+//extern Stepper motor_flow;
+//extern Stepper motor_valve_o2;
+//extern Stepper motor_valve_air;
 
 //this function bring a especific motor to a know position
 //return 0 is no succes o 1 if action was totaly complete
-bool DriverMotorMoveTo(int motor_id, int line_pos);
+bool DriverMotorMoveTo(int motor_id, float line_pos);
 
 //this function set the lin vel or rotation vel of motor
 //return 0 is no succes o 1 if action was totaly complete
-bool DriverMotorSetVel(int motor_id, int motor_vel);
+bool DriverMotorSetVel(int motor_id, float motor_vel);
 
-//this function initialize all develpment board GPIO
+//this function initialize all develpment board GPIO wich are not initialized by drivers
 //return 0 if there are some error else return 1
 bool PinInitialization(void);
 
@@ -32,7 +36,7 @@ bool DriverValveOpenTo(int valve_id, bool valve_position);
 long int SensorGetValue(int sensor_id);
 
 //return actual motor position
-int DriverMotorActualPos(int motor_id);
+float DriverMotorActualPos(int motor_id);
 
 //this function initialize all peripheral drivers
 //return 0 if there are some error else return 1
